--- website/src/components/sections/NewFeatures.tsx
@@ -1,169 +1,0 @@
-/* eslint-disable react/no-array-index-key */
-/* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
-/* eslint-disable jsx-a11y/click-events-have-key-events */
-import type { CSSProperties, FC } from 'react';
-import React, { useCallback, useState } from 'react';
-import Link from '@docusaurus/Link';
-import { animated } from 'react-spring';
-import Translate from '@docusaurus/Translate';
-import useBaseUrl from '@docusaurus/useBaseUrl';
-import { LazyLoadImage } from 'react-lazy-load-image-component';
-import styles from '../../css/features.module.scss';
-import ArrowAnim from '../ArrowAnim';
-
-const DashboardPlayground = () => (
-  <Link
-    className={styles.dashboardLink}
-    to="http://20.210.250.99:9000/user/login/"
-  >
-    <div className={styles.dashboard}>
-      <h3 className={styles.dashboardTitle}>Dashboard Playground</h3>
-
-      <div className={styles.dashboardAccount}>
-        <div>
-          username
-          {' '}
-          <span>admin</span>
-        </div>
-        <div>
-          passwd
-          {' '}
-          <span>admin</span>
-        </div>
-      </div>
-    </div>
-  </Link>
-);
-
-interface FeatProps {
-  style?: CSSProperties
-}
-
-type Feat = FC<FeatProps>
-
-const DashboardFeat: Feat = (props) => (
-  <animated.div className={styles.feat} {...props}>
-    <div className={styles.left}>
-      <section>
-        <h2>
-          <Translate id="features.component.easyDashboard.title">
-            Easy-to-use Dashboard
-          </Translate>
-        </h2>
-        <p>
-          <Translate id="features.component.easyDashboard.message">
-            The Apache APISIX Dashboard is designed to make it as easy as
-            possible for users to operate Apache APISIX through a frontend
-            interface. It’s opensource and ever evolving, feel free to
-            contribute.
-          </Translate>
-          <div className={styles.action}>
-            <Link to={useBaseUrl('downloads')} className="btn btn-download">
-              <Translate id="features.component.easyDashboard.downloadBtn">Downloads</Translate>
-            </Link>
-            <ArrowAnim />
-          </div>
-        </p>
-        <DashboardPlayground />
-      </section>
-    </div>
-
-    <div className={styles.right}>
-      <LazyLoadImage
-        src="https://static.apiseven.com/202202/apisix-dashboard.png"
-        alt="apisix-dashboard"
-      />
-    </div>
-  </animated.div>
-);
-
-const CodeFeat: Feat = (props) => (
-  <animated.div className={styles.feat} {...props}>
-    <div className={styles.left}>
-      <section>
-        <h2>
-          <Translate id="features.component.userFlexible.title">User Flexible</Translate>
-        </h2>
-        <p>
-          <Translate id="features.component.userFlexible.message">
-            The Apache APISIX dashboard is flexible to User demand,
-            providing option to create custom modules through code matching
-            your requirements, alongside the existing no-code toolchain.
-          </Translate>
-        </p>
-      </section>
-    </div>
-
-    <div className={styles.right}>
-      <LazyLoadImage
-        src="https://static.apiseven.com/202202/code-sample.png"
-        alt="code-snippet"
-      />
-    </div>
-  </animated.div>
-);
-
-const PluginFeat: Feat = (props) => (
-  <animated.div className={styles.feat} {...props}>
-    <div className={styles.left}>
-      <section>
-        <h2>
-          <Translate id="features.component.pluginised.title">Pluginised Workflow</Translate>
-        </h2>
-        <p>
-          <Translate id="features.component.pluginised.message">
-            No need to reinvent the wheel again and again. Use inbuilt
-            plugins to create high performance systems in tight deadlines.
-            For something custom, there is option of building custom
-            plugins.
-          </Translate>
-        </p>
-      </section>
-    </div>
-
-    <div className={styles.right}>
-      <LazyLoadImage
-        src="https://static.apiseven.com/202202/pluginised.png"
-        alt="plugin-workflow"
-      />
-    </div>
-  </animated.div>
-);
-
-const features: Feat[] = [
-  DashboardFeat,
-  CodeFeat,
-  PluginFeat,
-];
-
-const NewFeatures: FC = () => {
-  const [, setIdx] = useState(0);
-  const onClick = useCallback(() => setIdx((i) => (i + 1) % 3), []);
-
-  return (
-    <div className={styles.features}>
-
-      <section>
-        <h2 className="feat-head-desc"><Translate id="features.component.why.title">Why APISIX Gateway?</Translate></h2>
-        <h3 className="feat-head">
-          <Translate id="features.component.why.subtitle">
-            Reduce time fighting bugs, focus on designing world-class systems with API Gateway
-          </Translate>
-        </h3>
-        <p className="feat-desc">
-          <Translate id="features.component.why.message">
-            Apache APISIX is the first open-source API Gateway
-            that includes a built-in low-code Dashboard,
-            which offers a powerful and flexible UI for developers to use.
-          </Translate>
-        </p>
-      </section>
-
-      <div className={styles.container} onClick={onClick} role="article">
-        {features.map((Feat, i) => <Feat key={`feat${i}`} />)}
-      </div>
-    </div>
-  );
-};
-
-export default NewFeatures;
