+++ website/src/css/util.scss
@@ -3,7 +3,7 @@
 $breakpoints: (
   xs: 369px,
   sm: 736px,
-  md: 1024px,
+  md: 1025px,
   lg: 1500px,
 );
